rbd-mirror: notify MirroringWatcher on group create/delete
The group_replayer::BootstrapRequest was not notifying the MirroringWatcher on
creating or deleting the local group. When the rbd-mirror was restarted,
a mirror enabled group was deleted, and a new group with the same name
was mirror enabled, the ImapMap ended up with a non-empty peer_map and would
not send a release_group notification for the deleted group to the
InstanceReplayer.

Signed-off-by: N Balachandran <nithya.balachandran@ibm.com>